Refactor: 추출하기 트리거 함수 수정.
- SelectExtract, TriggerComplete 시트를 생성합니다.
- VBA 측에서 TriggerComplete 시트의 생성을 감지하여, SelectExtract  시트의 A1 셀의 값을 탐지합니다.
- 1초후 두 개의 시트를 삭제합니다.

+++ src/taskpane/utils/extractFileFunc.js
@@ -4,21 +4,40 @@ async function executeFunction(selectedOption) {
   try {
     Excel.run(async (context) => {
       const sheetName = "SelectExtract";
+      const completeSheetName = "TriggerComplete";
       let sheet = context.workbook.worksheets.getItemOrNullObject(sheetName);
+      let completeSheet =
+        context.workbook.worksheets.getItemOrNullObject(completeSheetName);
 
       await context.sync();
 
-      sheet.delete();
-      await context.sync();
+      if (!sheet.isNullObject) {
+        sheet.delete();
+        await context.sync();
+      }
 
       sheet = context.workbook.worksheets.add(sheetName);
       await context.sync();
 
-      sheet.getRange("A1").values = [[`${selectedOption}`]];
+      const targetRange = sheet.getRange("A1");
       await context.sync();
 
-      sheet.delete();
+      targetRange.values = [[`${selectedOption}`]];
       await context.sync();
+
+      if (!completeSheet.isNullObject) {
+        sheet.delete();
+        await context.sync();
+      }
+
+      completeSheet = context.workbook.worksheets.add(completeSheetName);
+      await context.sync();
+
+      setTimeout(async () => {
+        sheet.delete();
+        completeSheet.delete();
+        await context.sync();
+      }, 1000);
     });
   } catch (error) {
     updateState("setMessageList", {
